Q#1
The function q1 is solved by using Counting Sort Implementation which  takes an array of numbers as input and sorts it in descending order. It works by first finding the maximum and minimum values in the array to determine the range. It then creates a count array to keep track of the frequency of each number within this range. By iterating over the input array, it populates the count array. Finally, it reconstructs the original array in descending order based on the counts. This ensures that the array is sorted from the highest to the lowest value.