Q#2
Created a function that determines if a given string is a palindrome. The function should take a string as input, convert all characters to lowercase, and then check if the string reads the same from left to right as it does from right to left. If it does, the function returns true otherwise, it returns false.